Updated scratchattach files
